Moved RHF gradient code
