+++ forte2/gradients/__init__.py
@@ -1,4 +1,3 @@
-from .rhf import rhf_gradient
 from .utils import flat_to_atom_gradient, nuclear_repulsion_deriv
 
-__all__ = ["flat_to_atom_gradient", "nuclear_repulsion_deriv", "rhf_gradient"]
+__all__ = ["flat_to_atom_gradient", "nuclear_repulsion_deriv"]
